Move tests into main source file, so we can separate tests for individual source files more easily
